OpenPaaS-Suite/esn-frontend-admin#67 - Move inboxEmailer Directives To esn.inbox.libs
