+++ src/esn.inbox.libs/app/directives/inbox-emailer-group/inbox-emailer-group.directive.js
@@ -1,0 +1,15 @@
+'use strict';
+
+angular.module('esn.inbox.libs')
+
+  .directive('inboxEmailerGroup', function() {
+    return {
+      restrict: 'E',
+      replace: true,
+      scope: {
+        group: '=',
+        displayInline: '@?'
+      },
+      template: require('./inbox-emailer-group.pug')
+    };
+  });
