+++ src/esn.inbox.libs/app/directives/inbox-emailer/inbox-emailer.directive.js
@@ -1,0 +1,22 @@
+'use strict';
+
+angular.module('esn.inbox.libs')
+
+  .directive('inboxEmailer', function(session) {
+    return {
+      restrict: 'E',
+      replace: true,
+      controller: 'resolveEmailerController',
+      scope: {
+        emailer: '=',
+        hideEmail: '=?',
+        highlight: '@?'
+      },
+      template: require('./inbox-emailer.pug'),
+      link: function(scope) {
+        scope.$watch('emailer', function(emailer) {
+          scope.me = emailer && emailer.email && emailer.email === session.user.preferredEmail;
+        });
+      }
+    };
+  });
