+++ src/linagora.esn.unifiedinbox/app/directives/main.spec.js
@@ -844,71 +844,6 @@ describe('The linagora.esn.unifiedinbox Main module directives', function() {
 
   });
 
-  describe('The inboxEmailer directive', function() {
-
-    var session;
-
-    beforeEach(angular.mock.inject(function(_session_) {
-      session = _session_;
-    }));
-
-    it('should resolve the emailer', function() {
-      $scope.emailer = {
-        resolve: sinon.spy()
-      };
-
-      compileDirective('<inbox-emailer emailer="emailer"/>');
-
-      expect($scope.emailer.resolve).to.have.been.calledWith();
-    });
-
-    it('should resolve the emailer when it becomes available', function() {
-      compileDirective('<inbox-emailer emailer="emailer"/>');
-
-      $scope.emailer = {
-        resolve: sinon.spy()
-      };
-      $scope.$digest();
-
-      expect($scope.emailer.resolve).to.have.been.calledWith();
-    });
-
-    it('should not display the "me" message when the emailer is me', function() {
-      session.user = { preferredEmail: 'me@linagora.com' };
-      $scope.emailer = {
-        email: 'another-one@linagora.com',
-        resolve: angular.noop
-      };
-
-      compileDirective('<inbox-emailer emailer="emailer"/>');
-
-      expect(element.find('.me')).to.have.length(0);
-    });
-
-    it('should display the "me" message when the emailer is me', function() {
-      session.user = { preferredEmail: 'me@linagora.com' };
-      $scope.emailer = {
-        email: 'me@linagora.com',
-        resolve: angular.noop
-      };
-
-      compileDirective('<inbox-emailer emailer="emailer"/>');
-
-      expect(element.find('.me')).to.have.length(1);
-    });
-
-    it('should not display the email address if hide-email=true', function() {
-      $scope.emailer = {
-        email: 'me@linagora.com',
-        resolve: angular.noop
-      };
-
-      compileDirective('<inbox-emailer emailer="emailer" hide-email="true" />');
-
-      expect(element.find('.email')).to.have.length(0);
-    });
-  });
-
   describe('The inboxEmailerAvatar directive', function() {
 
     it('should resolve the emailer', function() {
@@ -1036,122 +971,6 @@ describe('The linagora.esn.unifiedinbox Main module directives', function() {
       compileDirective('<inbox-empty-container-message />');
 
       expect(element.isolateScope().containerTemplateUrl).to.equal('/unifiedinbox/views/partials/empty-messages/containers/inbox.html');
-    });
-
-  });
-
-  describe('The inboxEmailerDisplay directive', function() {
-
-    var email;
-
-    beforeEach(function() {
-      email = {
-        from: { name: 'Bob', email: 'bob@email', resolve: angular.noop },
-        to: [{ name: 'Alice', email: 'alice@email', resolve: angular.noop }],
-        cc: [{ name: 'Clark', email: 'clark@email', resolve: angular.noop }],
-        bcc: [{ name: 'John', email: 'john@email', resolve: angular.noop }]
-      };
-    });
-
-    it('should initialize by exposing scope attributes properly', function() {
-      $scope.email = email;
-      compileDirective('<inbox-emailer-display email="email" />');
-
-      var isolateScope = element.isolateScope();
-
-      expect(isolateScope.previewEmailer).to.deep.equal(email.to[0]);
-      expect(isolateScope.previewEmailerGroup).to.deep.equal('To');
-      expect(isolateScope.numberOfHiddenEmailer).to.equal(2);
-      expect(isolateScope.showMoreButton).to.equal(true);
-    });
-
-    it('should display the first "To" recipient', function() {
-      $scope.email = email;
-
-      compileDirective('<inbox-emailer-display email="email" />');
-
-      var isolateScope = element.isolateScope();
-
-      expect(isolateScope.previewEmailer.email).to.deep.equal('alice@email');
-      expect(isolateScope.previewEmailerGroup).to.deep.equal('To');
-    });
-
-    it('should display the first "CC" recipient', function() {
-      $scope.email = _.omit(email, 'to');
-
-      compileDirective('<inbox-emailer-display email="email" />');
-
-      var isolateScope = element.isolateScope();
-
-      expect(isolateScope.previewEmailer.email).to.deep.equal('clark@email');
-      expect(isolateScope.previewEmailerGroup).to.deep.equal('CC');
-    });
-
-    it('should display the first "BCC" recipient', function() {
-      $scope.email = _.omit(email, 'to', 'cc');
-
-      compileDirective('<inbox-emailer-display email="email" />');
-
-      var isolateScope = element.isolateScope();
-
-      expect(isolateScope.previewEmailer.email).to.deep.equal('john@email');
-      expect(isolateScope.previewEmailerGroup).to.deep.equal('BCC');
-    });
-
-    it('should be collapsed by default', function() {
-      $scope.email = email;
-      compileDirective('<inbox-emailer-display email="email" />');
-
-      expect(element.find('.recipients .collapsed, .more .collapsed').length).to.equal(2);
-      expect(element.find('.recipients .expanded, .more .expanded').length).to.equal(0);
-    });
-
-    it('should be expanded after a click on more button then collapsed when click again', function() {
-      $scope.email = email;
-      compileDirective('<inbox-emailer-display email="email" />');
-
-      element.find('.more').click();
-
-      expect(element.find('.recipients .collapsed, .more .collapsed').length).to.equal(0);
-      expect(element.find('.recipients .expanded, .more .expanded').length).to.equal(2);
-
-      element.find('.more').click();
-
-      expect(element.find('.recipients .collapsed, .more .collapsed').length).to.equal(2);
-      expect(element.find('.recipients .expanded, .more .expanded').length).to.equal(0);
-    });
-
-    it('should not show more button when there is only 1 recipient', function() {
-      $scope.email = {
-        from: { name: 'Bob', email: 'bob@email', resolve: angular.noop },
-        to: [{ name: 'Alice', email: 'alice@email', resolve: angular.noop }],
-        cc: []
-      };
-      compileDirective('<inbox-emailer-display email="email" />');
-
-      expect(element.find('.more').css('display')).to.equal('none');
-    });
-
-    it('should show both name and email if there is only 1 recipient and it is not current user', function() {
-      $scope.email = {
-        from: { name: 'Bob', email: 'bob@email', resolve: angular.noop },
-        to: [{ name: 'Alice', email: 'alice@email', resolve: angular.noop }],
-        cc: []
-      };
-
-      compileDirective('<inbox-emailer-display email="email" />');
-
-      expect(element.find('.to').html()).to.contain(email.to[0].name);
-      expect(element.find('.to').html()).to.contain(email.to[0].email);
-    });
-
-    it('should not display any recipients if there is no recipients', function() {
-      $scope.email = _.omit(email, 'to', 'cc', 'bcc');
-
-      compileDirective('<inbox-emailer-display email="email" />');
-
-      expect(element.find('.recipients .collapsed').length).to.equal(0);
-      expect(element.find('.recipients .expanded').length).to.equal(0);
     });
 
   });
